Split tools in a separate module to reduce dependencies
