+++ go.mod
@@ -3,17 +3,12 @@ module github.com/signalfx/sapm-proto
 go 1.13
 
 require (
-	github.com/client9/misspell v0.3.4
 	github.com/gogo/protobuf v1.3.2
 	github.com/golang/protobuf v1.4.3
-	github.com/golangci/golangci-lint v1.38.0
-	github.com/google/addlicense v0.0.0-20200906110928-a0294312aa76
+	github.com/google/go-cmp v0.5.4 // indirect
 	github.com/jaegertracing/jaeger v1.22.0
 	github.com/opentracing/opentracing-go v1.2.0 // indirect
-	github.com/ory/go-acc v0.2.6
-	github.com/pavius/impi v0.0.3
 	github.com/stretchr/testify v1.7.0
 	go.opencensus.io v0.23.0
 	google.golang.org/grpc v1.33.2
-	honnef.co/go/tools v0.1.2
 )
